feat: remove product to cart in page cart
