+++ src/app/features/account/pages/cart-page/cart-page.component.html
@@ -7,9 +7,7 @@
     <div class="flex w-1/4">summary</div>
 
     <div class="flex-1">
-      <app-list-cart-product
-        [productList]="productList()"
-      ></app-list-cart-product>
+      <app-list-cart-product></app-list-cart-product>
     </div>
   </div>
 </app-layout-content>
