+++ src/app/shared/components/ui/buttons/button-icon-small/button-icon-small.component.html
@@ -1,1 +1,6 @@
-<p-button [icon]="icon()" [rounded]="true" severity="secondary" />
+<p-button
+  [icon]="icon()"
+  [rounded]="true"
+  severity="secondary"
+  [ngClass]="{ 'has-background': hasBackground() }"
+/>
