+++ src/app/shared/components/ui/buttons/button-icon-small/button-icon-small.component.ts
@@ -1,12 +1,16 @@
-import { Component, InputSignal, input } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { Component, InputSignal, booleanAttribute, input } from '@angular/core';
 import { ButtonModule } from 'primeng/button';
 
 @Component({
 	selector: 'app-button-icon-small',
-	imports: [ButtonModule],
+	imports: [CommonModule, ButtonModule],
 	templateUrl: './button-icon-small.component.html',
 	styleUrl: './button-icon-small.component.scss',
 })
 export class ButtonIconSmallComponent {
 	public icon: InputSignal<string> = input.required<string>();
+	public hasBackground = input<boolean, unknown>(false, {
+		transform: booleanAttribute,
+	});
 }
