+++ src/app/shared/components/ui/card/card-product-cart/card-product-cart.component.html
@@ -4,6 +4,7 @@
   <div class="absolute -right-2 -top-2">
     <app-button-icon-small
       icon="pi pi-times"
+      hasBackground
       (click)="onDelete()"
     ></app-button-icon-small>
   </div>
