+++ src/app/shared/components/ui/card/card-product/card-product.component.html
@@ -9,6 +9,7 @@
     <div class="absolute right-4 top-4">
       <app-button-icon-small
         icon="pi pi-times"
+        hasBackground
         (click)="onClick()"
       ></app-button-icon-small>
     </div>
